refactor: content와 관련된 도메인을 따로 분리
