+++ back/src/main/java/com/baba/back/content/domain/content/ContentDate.java
@@ -1,4 +1,4 @@
-package com.baba.back.content.domain;
+package com.baba.back.content.domain.content;
 
 import com.baba.back.baby.domain.Birthday;
 import com.baba.back.content.exception.ContentDateBadRequestException;
