+++ back/src/test/java/com/baba/back/content/domain/ContentDateTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.baba.back.baby.domain.Birthday;
+import com.baba.back.content.domain.content.ContentDate;
 import com.baba.back.content.exception.ContentDateBadRequestException;
 import java.time.LocalDate;
 import java.util.stream.Stream;
